+++ examples/visualization/plot_channel_epochs_image.py
@@ -36,7 +36,7 @@ data_path = sample.data_path()
 # Set parameters
 raw_fname = data_path + '/MEG/sample/sample_audvis_filt-0-40_raw.fif'
 event_fname = data_path + '/MEG/sample/sample_audvis_filt-0-40_raw-eve.fif'
-event_id, tmin, tmax = 1, -0.2, 0.5
+event_id, tmin, tmax = 1, -0.2, 0.4
 
 # Setup for reading the raw data
 raw = io.read_raw_fif(raw_fname)
@@ -74,6 +74,6 @@ bad_pick = 98  # channel with no evoked response
 plt_times = np.linspace(0, .2, len(epochs))
 
 plt.close('all')
-mne.viz.plot_epochs_image(epochs, [good_pick, bad_pick], sigma=0.5, vmin=-100,
-                          vmax=250, colorbar=True, order=order_func,
+mne.viz.plot_epochs_image(epochs, [good_pick, bad_pick], sigma=.5,
+                          order=order_func, vmin=-250, vmax=250,
                           overlay_times=plt_times, show=True)