+++ mne/viz/ica.py
@@ -212,6 +212,7 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
     psd_args = dict() if psd_args is None else psd_args
     topomap_args = dict() if topomap_args is None else topomap_args
     image_args = dict() if image_args is None else image_args
+    image_args["ts_args"] = dict(truncate_xaxis=False)
     for d in (psd_args, topomap_args, image_args):
         if not isinstance(d, dict):
             raise ValueError('topomap_args, image_args and psd_args have to be'
@@ -222,14 +223,11 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
 
     # calculations
     # ------------
-    plot_line_at_zero = False
     if isinstance(inst, BaseRaw):
         # break up continuous signal into segments
         from ..epochs import _segment_raw
         inst = _segment_raw(inst, segment_length=2., verbose=False,
                             preload=True)
-    if inst.times[0] < 0. and inst.times[-1] > 0.:
-        plot_line_at_zero = True
 
     epochs_src = ica.get_sources(inst)
     ica_data = np.swapaxes(epochs_src.get_data()[:, picks, :], 0, 1)
@@ -273,15 +271,6 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
             [np.sqrt((d[d > 0] ** 2).mean(axis=0)) for d in diffs.T]]
         spectrum_std = np.array(spectrum_std) * num_std
 
-        # erp std
-        if plot_std:
-            erp = ica_data[idx].mean(axis=0)
-            diffs = ica_data[idx] - erp
-            erp_std = [
-                [np.sqrt((d[d < 0] ** 2).mean(axis=0)) for d in diffs.T],
-                [np.sqrt((d[d > 0] ** 2).mean(axis=0)) for d in diffs.T]]
-            erp_std = np.array(erp_std) * num_std
-
         # epoch variance
         epoch_var = np.var(ica_data[idx], axis=1)
 
@@ -291,7 +280,7 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
         _plot_ica_topomap(ica, pick, show=False, axes=axes[0], **topomap_args)
 
         # image and erp
-        plot_epochs_image(epochs_src, picks=pick, axes=axes[1:3],
+        plot_epochs_image(epochs_src, picks=pick, axes=axes[1:3], combine=None,
                           colorbar=False, show=False, **image_args)
 
         # spectrum
@@ -299,10 +288,10 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
         if plot_std:
             axes[3].fill_between(freqs, psds_mean - spectrum_std[0],
                                  psds_mean + spectrum_std[1],
-                                 color='k', alpha=.15)
+                                 color='k', alpha=.2)
         if plot_lowpass_edge:
             axes[3].axvline(inst.info['lowpass'], lw=2, linestyle='--',
-                            color='k', alpha=0.15)
+                            color='k', alpha=0.2)
 
         # epoch variance
         axes[4].scatter(range(len(epoch_var)), epoch_var, alpha=0.5,
@@ -315,27 +304,14 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
         set_title_and_labels(axes[1], 'epochs image and ERP/ERF', [], 'Epochs')
 
         # erp
-        set_title_and_labels(axes[2], [], 'time', 'AU')
-        # line color and std
-        axes[2].lines[0].set_color('k')
-        if plot_std:
-            erp_xdata = axes[2].lines[0].get_data()[0]
-            axes[2].fill_between(erp_xdata, erp - erp_std[0],
-                                 erp + erp_std[1], color='k', alpha=.15)
-            axes[2].autoscale(enable=True, axis='y')
-            axes[2].axis('auto')
-            axes[2].set_xlim(erp_xdata[[0, -1]])
+        set_title_and_labels(axes[2], [], 'time', 'AU\n')
+        axes[2].spines["right"].set_color('k')
+        axes[2].set_xlim(epochs_src.times[[0, -1]])
         # remove half of yticks if more than 5
         yt = axes[2].get_yticks()
         if len(yt) > 5:
             yt = yt[::2]
             axes[2].yaxis.set_ticks(yt)
-
-        if not plot_line_at_zero:
-            xlims = [1e3 * inst.times[0], 1e3 * inst.times[-1]]
-            for k, ax in enumerate(axes[1:3]):
-                ax.lines[k].remove()
-                ax.set_xlim(xlims)
 
         # remove xticks - erp plot shows xticks for both image and erp plot
         axes[1].xaxis.set_ticks([])
@@ -351,6 +327,7 @@ def plot_ica_properties(ica, inst, picks=None, axes=None, dB=True,
         ylim = axes[3].get_ylim()
         air = np.diff(ylim)[0] * 0.1
         axes[3].set_ylim(ylim[0] - air, ylim[1] + air)
+        axes[1].axhline(0, color='k', linewidth=.5)
 
         # epoch variance
         set_title_and_labels(axes[4], 'epochs variance', 'epoch', 'AU')